backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
mysql-connector-python==8.3.0
python-dotenv==1.0.0
pydantic==2.5.3
python-multipart==0.0.6
mysqlclient==2.1.1
pdfplumber==0.10.3
requests==2.31.0
python-docx==1.1.0
reportlab==4.0.7
qdrant-client
email-validator
sentence-transformers